functions/status_updater_function/requirements.txt
google-cloud-bigquery
functions-framework